+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## HEAD
 
+* Add `:force` option to `infer_extension` plugin for always replacing the current extension (@jrochkind)
+
 * Add `:public` option to `S3#initialize` for enabling public uploads (@janko-m)
 
 * Add ability to specify a custom `:signer` for `Shrine::Storage::S3#url` (@janko-m)